### 2019-07-26 #41 @whedon
[PRE REVIEW]: Markov Chain-Monte Carlo Methods for Linear Algebra using Julia v.1.0
**Submitting author:** @oaefbsc (<a href="http://orcid.org/0000-0003-1451-1285">Oscar Alejandro Esquivel-Flores</a>)
**Repository:** <a href="https://github.com/oaefbsc/Juliacon2019" target ="_blank">https://github.com/oaefbsc/Juliacon2019</a>

Thanks for submitting your paper to JuliaCon @oaefbsc. The JuliaCon editor @matbesancon, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@oaefbsc if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
