+++ paper/paper.yml
@@ -9,12 +9,8 @@ authors:
     affiliation: 1
   - name: H\'{e}ctor Ben\'{i}tez P\'{e}rez
     affiliation: 1
-  - name:
-    affiliation: 
 affiliations:
   - name: Universidad Nacional Autónoma de México, IIMAS-DISCA.
     index: 1
-  - name: 
-    index:
 date: 23 July 2019
 bibliography: ref.bib
