### 2019-09-11 #41 @oaefbsc
Apreciated Mahtie,

I have been a lot of problems, personal and technical, about to solve the
problem with my Juliacon2019 paper.
I would like to kwon if is possible to do somethin about review it and
include in proceedings.

I will appreciate your answer.

Regards,
Oscar Esquivel

El mar., 20 ago. 2019 a las 3:23, Mathieu Besançon (<
notifications@github.com>) escribió:

> ping @oaefbsc <https://github.com/oaefbsc>
>
