### 2020-07-27 #60 @whedon
[REVIEW]: Markov Chain-Monte Carlo Methods for Linear Algebra using Julia v.1.0
**Submitting author:** @oaefbsc (<a href="http://orcid.org/0000-0003-1451-1285">Oscar Alejandro Esquivel-Flores</a>)
**Repository:** <a href="https://github.com/oaefbsc/Juliacon2019" target ="_blank">https://github.com/oaefbsc/Juliacon2019</a>
**Version:**
