### 2021-01-16 #60 @oaefbsc
Please!!

I would like to see this paper published.

Thanks for working on the article review process.

Regards,
Oscar Esquivel

El El jue, 14 de enero de 2021 a la(s) 9:49, Valentin Churavy <
notifications@github.com> escribió:

> @oaefbsc <https://github.com/oaefbsc> sorry for all the delays. Are you
> still interested in seeing this paper published?
>
